core: fix three latent UB/correctness bugs
- logging: handle vsnprintf negative-return error case so a -1 doesn't
  convert to SIZE_MAX when passed to log_write.
- stats: guard minmax_log against divide-by-zero when count == 0.
- net_compute_checksum: avoid signed-int overflow on AVR (16-bit int)
  by promoting the multiplication to unsigned.

+++ src/lib/core/net_compute_checksum.c
@@ -22,7 +22,9 @@ uint8_t net_compute_checksum(const unsigned char *buf, int size) {
   int i;
   uint8_t cksum = 0x67;
   for (i = 0; i < size; i++) {
-    cksum = cksum * 131 + buf[i];
+    // Cast to unsigned to avoid signed-int overflow on AVR, where int
+    // is 16 bits and 255*131+255 = 33660 exceeds INT16_MAX.
+    cksum = (uint8_t)((unsigned)cksum * 131U + buf[i]);
   }
   return cksum;
 }
